Added various mechanics
This commit is very big, but in brief:
Added a movement tracking mechanic
Added gravity
Added platforms
Added Lightning strike mechanic
